requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
boto3==1.26.76
botocore==1.29.76
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
pydantic==1.10.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
